requirements.txt
requests == 2.22.0
beautifulsoup4 == 4.8.2
rutermextract == 0.3
PyQt5 == 5.14.1
python-docx